fix(macOS与翻译): 收尾 Vulkan 构建与翻译设置问题
- 调整 macOS 构建与通知实现，清理剩余编译警告\n- 修复翻译弹窗交互与出站自动翻译默认模式\n- 补充并修正相关测试，统一忽略本地 CLAUDE.md

+++ src/test/translate_boundary_test.cpp
@@ -305,7 +305,7 @@ TEST(TranslateBoundary, TranslationTextLengthBoundaries)
 
 	// 最大长度文本
 	char aMaxText[MAX_LINE_LENGTH];
-	for(int i = 0; i < MAX_LINE_LENGTH - 1; i++)
+	for(size_t i = 0; i < MAX_LINE_LENGTH - 1; i++)
 	{
 		aMaxText[i] = 'a';
 	}
